// src/app/models/ner/pr_summaryReport.ts
export class PrSummaryReport {
  prYear: string = '';
  deptCode: string = '';
  deptSymbol: string = '';
  division: string = '';
  prStatus: string = '';
  prPending: number = 0;
  prCanceled: number = 0;
  prChecked: number = 0;
  prApproved: number = 0;
  prCompleted: number = 0;
  prAll: number = 0;
  prKpi: number = 0;
  avgKpi: number = 0;
}